# Faster logical operator construction