--- a/qldpc/codes.py
+++ b/qldpc/codes.py
@@ -681,8 +681,8 @@ class CSSCode(QuditCode):
         procedure described in arXiv:0903.5256, slightly modified and generalized for qudits.
         """
         # memoize manually because other methods may modify the logical operators computed here
-        # if self._logical_ops is not None:
-        #     return self._logical_ops
+        if self._logical_ops is not None:
+            return self._logical_ops
 
         num_qudits = self.num_qudits
         dimension = self.dimension
@@ -769,8 +769,8 @@ class CSSCode(QuditCode):
         procedure described in arXiv:0903.5256, slightly modified and generalized for qudits.
         """
         # memoize manually because other methods may modify the logical operators computed here
-        # if self._logical_ops is not None:
-        #     return self._logical_ops
+        if self._logical_ops is not None:
+            return self._logical_ops
 
         # collect logical operators sequentially
         logicals_x: list[galois.FieldArray] = []