--- a/qldpc/codes.py
+++ b/qldpc/codes.py
@@ -677,8 +677,8 @@ class CSSCode(QuditCode):
         by this method are the unit shift and phase operators that generate all logical X-type and
         Z-type qudit operators.
 
-        Logical operators are identified using the symplectic Gram-Schmidt orthogonalization
-        procedure described in arXiv:0903.5256, slightly modified and generalized for qudits.
+        Logical operators are constructed using the method described in Section 4.1 of Gottesman's
+        thesis (arXiv:9705052), slightly modified and generalized for qudits.
         """
         # memoize manually because other methods may modify the logical operators computed here
         if self._logical_ops is not None: