--- a/qldpc/codes_test.py
+++ b/qldpc/codes_test.py
@@ -110,13 +110,21 @@ def test_CSS_code() -> None:
 
 def test_logical_ops() -> None:
     """Logical operator construction."""
-    code = codes.HGPCode(codes.ClassicalCode.random(4, 2, field=3))
+    code = codes.HGPCode(codes.ClassicalCode.random(4, 2, field=2))
     code.get_random_logical_op(codes.Pauli.X, ensure_nontrivial=False)
     code.get_random_logical_op(codes.Pauli.X, ensure_nontrivial=True)
-    ops = code.get_logical_ops()
-    for qq in range(code.dimension):
-        for rr in range(qq + 1):
-            assert ops[0, qq, :] @ ops[1, rr, :] == int(qq == rr)
+
+    logicals = code.get_logical_ops()
+    logicals_x, logicals_z = logicals[0], logicals[1]
+    assert np.array_equal(logicals_x @ logicals_z.T, np.eye(code.dimension, dtype=int))
+    assert not np.any(code.code_z.matrix @ logicals_x.T)
+    assert not np.any(code.code_x.matrix @ logicals_z.T)
+
+    logicals = code.new_get_logical_ops()
+    logicals_x, logicals_z = logicals[0], logicals[1]
+    assert np.array_equal(logicals_x @ logicals_z.T, np.eye(code.dimension, dtype=int))
+    assert not np.any(code.code_z.matrix @ logicals_x.T)
+    assert not np.any(code.code_x.matrix @ logicals_z.T)
 
 
 def test_deformations(num_qudits: int = 5, num_checks: int = 3, field: int = 3) -> None: